+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `InvitationCreated` event, dispatched whenever an invitation is created, as the hook for sending application-defined notifications.
+
 ## [0.0.2] - 2026-09-16
 
 ### Fixed